// bookSprite_FE/src/app/models/GoogleBook.ts
export class GoogleBook {
  id: number;
  title: string;
  authors: string;
  publisher: string;
  publishedDate: string;
  description: string;
  averageRating: number;
  language: string;
  pageCount: number;
  printType: string;
  ratingsCount: number;
  smallThumbnail: string;
  thumbnail: string;
  rating: number;
}